Modify all_check.py Script
Rename all_check.py to pc_health_check.py: to clarify the
the function of the script.

Modify all the functions in all_check.py.
Added extra functions to improve the functionality of the script.

Modify the script to save the output of the script to a text file
instead printing it to the terminal.

Added current_date module: thi module help which keep track of
current date and make sure a current date is not outputed twice in
the same day except when the script is run for the first time or in
a new date.